Can now edit image mask for threshold image (#1)
Image mask can be edited from all images. When edited on original image
the mask is replaced, on others the mask is merged.

MaskPainterDialog is now integrated with OCRGraphicsView rather than
MainWindow.

MaskPainterGraphicsView::drawCircle() tried to draw out of bounds, now
fixed.

Added ImageUtility::mergeAlpha() for merging an alpha image with another
image. It treats the alpha channel as binary and multiplies them.