Added MultiscaleOCR for performing OCR at multiple scales. (#2)
Results are currently filtered in OCRGraphicsView but should be filtered
in MultiscaleOCR.
They are filtered using ascender, but font metric calculation needs
improved.

Added CharProperty which provides where the top and bottom of a char is
expected to lie. This is then used to help calculate font metrics in
FontMetric.