FontMetric now calculates median and ascender (#2)
Still need to add percentage guesses when one cannot be calculated.
The capital from i and j are still not calculated correctly (as the dot
is ignored).

Some of the text I have been testing with has a backdrop/3D-ish effect
which is resulting in small connected components near the text. This
messes up the font metric calculations.

All things considered the results seem to be fairly good.