Added OCR Correction tab (#8)
First you preprocess an image in the "Image Preprocessing" tab then click the OCR button. Then the results are displayed in the "OCR Correction" tab. Correction not yet implemented.

OCRGraphicsView for some reason initially displays super small.

Sometimes the MultiscaleOCR crashes (seems to be at the start of FontMetric constructor).